Setup a signal that emits when the properties of a `has_properties` class change
Currently there is a signal that emits each time a property changes.
This means that for a single class, many signals may be emitted in a very short amount of time.
To a plugin that simply wants to know that properties have changed, but don't really care
beyond that, having to handle all of them at once is excessive.

Thus, this emits a generic "they've changed" message.