Check needed plugins are available on showfile load
It is possible to create showfiles that use plugins that another user or
computer may not have installed. This causes a cascade of errors if such a
showfile is loaded.

This commit implements a check and a dialog that (for showfiles with the
necessary metadata) warns if needed plugins are not installed, or are but are
not enabled.